refactor(gateway): extract signal_ingestion.go from server.go
Phase 3b file split: moves RegisterAdapter, createAdapterHandler,
handleBatchRequest, processSingleSignal, processMultiSignalBatch,
maxRequestBodySize, readParseValidateSignal, handleProcessingError,
sendSuccessResponse, validateScope, ProcessSignal, handleDuplicateSignal,
and createRemediationRequestCRD (the HTTP ingestion + processing pipeline)
into their own file. Pure structural move: 154 Ginkgo specs still pass.
This completes the server.go split (2552 -> 633 lines across 6 files).
